requirements.txt
Flask
flask-cors
chromadb
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
PyMuPDF
python-dotenv
gunicorn
werkzeug
redis
celery
faiss-cpu